simplify unified checkpoint cases
