+++ llm/llama/tests/utils.py
@@ -22,11 +22,13 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args = copy.deepcopy(pretrain_arguments)
     train_args["tensor_parallel_degree"] = 8
     train_args["pipeline_parallel_degree"] = 1
+    train_args["pretrain_arguments_name"] = "TP8"
     configs["TP8"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
     train_args["tensor_parallel_degree"] = 4
     train_args["pipeline_parallel_degree"] = 2
+    train_args["pretrain_arguments_name"] = "TP4PP2"
     configs["TP4PP2"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -34,6 +36,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["pipeline_parallel_degree"] = 1
     train_args["sharding"] = ""
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 2
+    train_args["pretrain_arguments_name"] = "TP4DP2"
     configs["TP4DP2"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -41,11 +44,13 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["pipeline_parallel_degree"] = 1
     train_args["sharding"] = "stage1"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 2
+    train_args["pretrain_arguments_name"] = "TP4Sharding2"
     configs["TP4Sharding2"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
     train_args["tensor_parallel_degree"] = 2
     train_args["pipeline_parallel_degree"] = 4
+    train_args["pretrain_arguments_name"] = "TP2PP4"
     configs["TP2PP4"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -53,11 +58,13 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["pipeline_parallel_degree"] = 1
     train_args["sharding"] = "stage1"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 4
+    train_args["pretrain_arguments_name"] = "TP2Sharding4"
     configs["TP2Sharding4"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
     train_args["tensor_parallel_degree"] = 1
     train_args["pipeline_parallel_degree"] = 8
+    train_args["pretrain_arguments_name"] = "PP8"
     configs["PP8"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -65,6 +72,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["pipeline_parallel_degree"] = 4
     train_args["sharding"] = ""
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 2
+    train_args["pretrain_arguments_name"] = "PP4DP2"
     configs["PP4DP2"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -72,6 +80,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["pipeline_parallel_degree"] = 4
     train_args["sharding"] = "stage1"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 2
+    train_args["pretrain_arguments_name"] = "PP4Sharding2"
     configs["PP4Sharding2"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -79,6 +88,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["pipeline_parallel_degree"] = 1
     train_args["sharding"] = "stage1"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 8
+    train_args["pretrain_arguments_name"] = "Sharding8S1"
     configs["Sharding8S1"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -86,6 +96,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["pipeline_parallel_degree"] = 1
     train_args["sharding"] = "stage2"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 8
+    train_args["pretrain_arguments_name"] = "Sharding8S2"
     configs["Sharding8S2"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -94,6 +105,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["sharding_parallel_degree"] = 4
     train_args["sharding"] = "stage1"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 8
+    train_args["pretrain_arguments_name"] = "Sharding4S1DP2"
     configs["Sharding4S1DP2"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -102,6 +114,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["sharding_parallel_degree"] = 4
     train_args["sharding"] = "stage2"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 8
+    train_args["pretrain_arguments_name"] = "Sharding4S2DP2"
     configs["Sharding4S2DP2"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -110,6 +123,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["sharding_parallel_degree"] = 2
     train_args["sharding"] = "stage1"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 8
+    train_args["pretrain_arguments_name"] = "Sharding2S1DP4"
     configs["Sharding2S1DP4"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -118,6 +132,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["sharding_parallel_degree"] = 2
     train_args["sharding"] = "stage2"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 8
+    train_args["pretrain_arguments_name"] = "Sharding2S2DP4"
     configs["Sharding2S2DP4"] = train_args
 
     train_args = copy.deepcopy(pretrain_arguments)
@@ -126,6 +141,7 @@ def get_pretrain_arguments(pretrain_arguments):
     train_args["sharding_parallel_degree"] = 1
     train_args["sharding"] = "stage2"
     train_args["gradient_accumulation_steps"] = train_args["gradient_accumulation_steps"] // 8
+    train_args["pretrain_arguments_name"] = "DP8"
     configs["DP8"] = train_args
 
     return configs
